Saca Duty Cycle

+++ Core/Inc/BCU/Communication/SPI.hpp
@@ -13,40 +13,40 @@ class SPI {
     StateMachine::state_id slave_nested_state{
         Shared::State::SharedStateMachine::NestedState::Idle};
 
-    double velocity_reference{0.0};
-    double velocity_error{0.0};
-    double u_current_measurement{0.0};
-    double v_current_measurement{0.0};
-    double w_current_measurement{0.0};
-    double electrical_angle{0.0};
-    double d_current_reference{0.0};
-    double d_current_measurement{0.0};
-    double d_current_error{0.0};
-    double q_current_reference{0.0};
-    double q_current_measurement{0.0};
-    double q_current_error{0.0};
-    double three_phase_unbalance{0.0};
-    double d_target_voltage{0.0};
-    double q_target_voltage{0.0};
-    double u_target_voltage{0.0};
-    double v_target_voltage{0.0};
-    double w_target_voltage{0.0};
-    double u_output_voltage{0.0};
-    double v_output_voltage{0.0};
-    double w_output_voltage{0.0};
-    double u_duty_cycle{0.0};
-    double v_duty_cycle{0.0};
-    double w_duty_cycle{0.0};
-    double angular_velocity{0.0};
-    std::array<double, 3> position{{0.0, 0.0, 0.0}};
-    std::array<double, 3> velocity{{0.0, 0.0, 0.0}};
-    std::array<double, 3> acceleration{{0.0, 0.0, 0.0}};
+    float velocity_reference{0.0};
+    float velocity_error{0.0};
+    float u_current_measurement{0.0};
+    float v_current_measurement{0.0};
+    float w_current_measurement{0.0};
+    float electrical_angle{0.0};
+    float d_current_reference{0.0};
+    float d_current_measurement{0.0};
+    float d_current_error{0.0};
+    float q_current_reference{0.0};
+    float q_current_measurement{0.0};
+    float q_current_error{0.0};
+    float three_phase_unbalance{0.0};
+    float d_target_voltage{0.0};
+    float q_target_voltage{0.0};
+    float u_target_voltage{0.0};
+    float v_target_voltage{0.0};
+    float w_target_voltage{0.0};
+    float u_output_voltage{0.0};
+    float v_output_voltage{0.0};
+    float w_output_voltage{0.0};
+    float u_duty_cycle{0.0};
+    float v_duty_cycle{0.0};
+    float w_duty_cycle{0.0};
+    float angular_velocity{0.0};
+    std::array<float, 3> position{{0.0, 0.0, 0.0}};
+    std::array<float, 3> velocity{{0.0, 0.0, 0.0}};
+    std::array<float, 3> acceleration{{0.0, 0.0, 0.0}};
     std::array<Shared::Communication::Direction, 3> direction{
         {Shared::Communication::Direction::FORWARD,
          Shared::Communication::Direction::FORWARD,
          Shared::Communication::Direction::FORWARD}};
-    double average_position{0.0};
-    double max_velocity{0.0};
+    float average_position{0.0};
+    float max_velocity{0.0};
     bool is_detecting{false};
     std::array<float, 4> dc_link_voltage{{0.0f, 0.0f, 0.0f, 0.0f}};
     std::array<std::array<float, 3>, 4> motor_phase_current{
@@ -60,13 +60,13 @@ class SPI {
         {PinState::OFF, PinState::OFF, PinState::OFF, PinState::OFF}};
 
    private:
-    double requested_velocity_reference{0.0};
-    double requested_d_current_reference{0.0};
-    double requested_q_current_reference{0.0};
-    double requested_angular_velocity{0.0};
-    double requested_duty_cycle_u{0.0};
-    double requested_duty_cycle_v{0.0};
-    double requested_duty_cycle_w{0.0};
+    float requested_velocity_reference{0.0};
+    float requested_d_current_reference{0.0};
+    float requested_q_current_reference{0.0};
+    float requested_angular_velocity{0.0};
+    float requested_duty_cycle_u{0.0};
+    float requested_duty_cycle_v{0.0};
+    float requested_duty_cycle_w{0.0};
     float requested_dc_link_voltage{0.0f};
     uint32_t requested_commutation_frequency_hz{0};
     uint32_t requested_commutation_dead_time_ns{0};
@@ -110,14 +110,14 @@ class SPI {
 
     void enable_booster();
 
-    void start_velocity_control(double velocity_reference);
-    void start_current_control(double d_current_reference,
-                               double q_current_reference);
-    void start_emulated_movement(double d_current_reference,
-                                 double q_current_reference,
-                                 double angular_velocity);
-    void start_test_pwm(double duty_cycle_u, double duty_cycle_v,
-                        double duty_cycle_w);
+    void start_velocity_control(float velocity_reference);
+    void start_current_control(float d_current_reference,
+                               float q_current_reference);
+    void start_emulated_movement(float d_current_reference,
+                                 float q_current_reference,
+                                 float angular_velocity);
+    void start_test_pwm(float duty_cycle_u, float duty_cycle_v,
+                        float duty_cycle_w);
 
     void stop_control();
